allow import of top-level seefarmer module is seefarmer.yaml doesn't exist
