+++ seedfarmer/mgmt/module_info.py
@@ -17,7 +17,7 @@ import os
 from enum import Enum
 from typing import Any, Dict, List, Optional, Tuple
 
-from seedfarmer import OPS_ROOT, PROJECT
+from seedfarmer.config import OPS_ROOT, PROJECT
 from seedfarmer.services import _secrets_manager as secrets
 from seedfarmer.services import _ssm as store
 from seedfarmer.utils import generate_hash
